refactor: products view and serializer
